Allow fetchAllAirbyteJobs to take a parital argument for the job list body Move/refactor fetching airbyte jobs to an exported util in airbyte api
